Removed the alias keys function
